Move generateTableCSV to common folder and reuse it.
