+++ packages/common/src/utils/generateReports/generateCLICookieTableCSV.ts
@@ -19,7 +19,7 @@
 import { type CookieTableData, sanitizeCsvRecord } from '@google-psat/common';
 import { I18n } from '@google-psat/i18n';
 
-const COOKIES_TABLE_DATA_HEADER_EXTENSION = [
+const COOKIES_TABLE_DATA_HEADER_CLI = [
   () => I18n.getMessage('name'),
   () => I18n.getMessage('scope'),
   () => I18n.getMessage('domain'),
@@ -34,11 +34,12 @@ const COOKIES_TABLE_DATA_HEADER_EXTENSION = [
   () => I18n.getMessage('expires'),
   () => I18n.getMessage('issues'),
   () => I18n.getMessage('gdpr'),
-  () => I18n.getMessage('priority'),
-  () => I18n.getMessage('size'),
 ];
 
-const generateExtensionCookieTableCSV = (cookies: CookieTableData[]): Blob => {
+const generateCLICookieTableCSV = (
+  cookies: CookieTableData[],
+  returnString = false
+): Blob | string => {
   let cookieRecords = '';
 
   for (const cookie of cookies) {
@@ -50,7 +51,8 @@ const generateExtensionCookieTableCSV = (cookies: CookieTableData[]): Blob => {
         : I18n.getMessage('thirdParty'),
       cookie.parsedCookie.domain || ' ',
       cookie.parsedCookie.partitionKey || ' ',
-      cookie.parsedCookie.samesite,
+      //@ts-ignore
+      cookie.parsedCookie?.sameSite ?? cookie.parsedCookie?.samesite,
       I18n.getMessage(
         cookie.analytics?.category?.toLowerCase() || 'uncategorized'
       ),
@@ -63,23 +65,27 @@ const generateExtensionCookieTableCSV = (cookies: CookieTableData[]): Blob => {
         : I18n.getMessage('no'),
       cookie.parsedCookie.value,
       cookie.parsedCookie.path,
-      cookie.parsedCookie.expires === 'Session'
-        ? I18n.getMessage('session')
-        : cookie.parsedCookie.expires,
+      cookie.parsedCookie.expires,
       cookie.isBlocked ? I18n.getMessage('yes') : I18n.getMessage('no'),
       cookie.analytics?.gdprUrl || 'NA',
-      I18n.getMessage((cookie.parsedCookie?.priority || '').toLowerCase()),
-      cookie.parsedCookie.size?.toString(),
     ].map(sanitizeCsvRecord);
 
     cookieRecords += recordsArray.join(',') + '\r\n';
   }
 
+  if (returnString) {
+    return (
+      COOKIES_TABLE_DATA_HEADER_CLI.map((header) => header()).join(',') +
+      '\r\n' +
+      cookieRecords
+    );
+  }
+
   return new Blob([
-    COOKIES_TABLE_DATA_HEADER_EXTENSION.map((header) => header()).join(',') +
+    COOKIES_TABLE_DATA_HEADER_CLI.map((header) => header()).join(',') +
       '\r\n' +
       cookieRecords,
   ]);
 };
 
-export default generateExtensionCookieTableCSV;
+export default generateCLICookieTableCSV;
