+++ packages/design-system/src/components/table/utils/index.ts
@@ -13,7 +13,4 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
-export { default as generateCLICookieTableCSV } from './generateCLICookieTableCSV';
-export { default as generateExtensionCookieTableCSV } from './generatExtensionCookietableCSV';
 export * from './precalculatedFiltersUtils';
